#1764 fixed pagination, added cert select as component
